+++ learn/matcher.ts
@@ -1237,9 +1237,57 @@ export class Matcher {
 
     static age(c: Character): number | null {
         const rawAge = Matcher.getTagValue(TagId.Age, c);
-        const age = ((rawAge) && (rawAge.string)) ? parseInt(rawAge.string, 10) : null;
+
+        if (!rawAge || !rawAge.string) {
+            return null;
+        }
+
+        const ageStr = rawAge.string.toLowerCase().trim();
+
+        if ((ageStr.indexOf('shota') >= 0) || (ageStr.indexOf('loli') >= 0) || (ageStr.indexOf('lolli') >= 0)) {
+            return 10;
+        }
+
+        const age = parseInt(rawAge.string, 10);
 
         return age && !Number.isNaN(age) && Number.isFinite(age) ? age : null;
+    }
+
+    static apparentAge(c: Character): { min: number, max: number } | null {
+        const rawAge = Matcher.getTagValue(TagId.ApparentAge, c);
+
+        if ((!rawAge) || (!rawAge.string)) {
+            return null;
+        }
+
+        const ageStr = rawAge.string.trim().toLowerCase();
+
+        if (ageStr === '') {
+            return null;
+        }
+
+        // '18'
+        if (/^[0-9]+$/.exec(ageStr)) {
+            const val = parseInt(rawAge.string, 10);
+
+            return { min: val, max: val };
+        }
+
+        // '18-22'
+        const rangeMatch = ageStr.match(/^([0-9]+)-([0-9]+)$/);
+
+        if (rangeMatch) {
+            const v1 = parseInt(rangeMatch[1], 10);
+            const v2 = parseInt(rangeMatch[2], 10);
+
+            return { min: Math.min(v1, v2), max: Math.max(v1, v2) };
+        }
+
+        if ((ageStr.indexOf('shota') >= 0) || (ageStr.indexOf('loli') >= 0) || (ageStr.indexOf('lolli') >= 0)) {
+            return { min: 10, max: 10 };
+        }
+
+        return null;
     }
 
     static calculateSearchScoreForMatch(